one of the areas in NV goes up to 10 billion. use int64.

+++ District2.h
@@ -85,7 +85,7 @@ public:
 
 	static void step();
 
-	inline void _addFirstToCenter( double x, double y, uint32_t tarea ) {
+	inline void _addFirstToCenter( double x, double y, uint64_t tarea ) {
 	    if ( tarea == 0 ) {
 		distx = wx = x;
 		disty = wy = y;
